Add Handle Errors
