+++ src/main/java/NobitaException.java
@@ -1,0 +1,10 @@
+public class NobitaException extends Exception {
+    public NobitaException(String errorMessage) {
+        super(errorMessage);
+    }
+
+    @Override
+    public String toString() {
+        return "☹ OOPS!!! " + super.toString();
+    }
+}
